Make healthPoint inherite from previous level's

+++ app/src/main/java/com/kokayapp/rocket9/Rocket.java
@@ -26,7 +26,7 @@ public class Rocket extends MovableObject {
         velocityX = 3;
         maxVelocity = 6;
         acceleration = 6;
-        this.healthPoint = healthPoint;
+        this.healthPoint = (float) healthPoint;
         maxHealthPoint = MAX_HEALTH_POINT;
         setLevel(level);
         velocityY = 0;
